+++ src/main/java/com/fairing/fairplay/statistics/repository/dailystats/DailyStatsCustomRepository.java
@@ -1,7 +1,7 @@
 package com.fairing.fairplay.statistics.repository.dailystats;
 
 
-import com.fairing.fairplay.statistics.entity.EventDailyStatistics;
+import com.fairing.fairplay.statistics.entity.reservation.EventDailyStatistics;
 import java.time.LocalDate;
 import java.util.List;
