+++ src/main/java/com/fairing/fairplay/statistics/repository/dailystats/EventDailyStatisticsRepository.java
@@ -1,6 +1,6 @@
 package com.fairing.fairplay.statistics.repository.dailystats;
 
-import com.fairing.fairplay.statistics.entity.EventDailyStatistics;
+import com.fairing.fairplay.statistics.entity.reservation.EventDailyStatistics;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.time.LocalDate;
