+++ src/main/java/com/fairing/fairplay/statistics/repository/hourlystats/EventHourlyStatisticsRepository.java
@@ -1,6 +1,6 @@
 package com.fairing.fairplay.statistics.repository.hourlystats;
 
-import com.fairing.fairplay.statistics.entity.EventHourlyStatistics;
+import com.fairing.fairplay.statistics.entity.reservation.EventHourlyStatistics;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EventHourlyStatisticsRepository extends JpaRepository<EventHourlyStatistics, Long>,HourlyStatsCustomRepository {}
