+++ src/main/java/com/fairing/fairplay/statistics/repository/ticketstats/TicketStatsCustomRepository.java
@@ -1,6 +1,6 @@
 package com.fairing.fairplay.statistics.repository.ticketstats;
 
-import com.fairing.fairplay.statistics.entity.EventTicketStatistics;
+import com.fairing.fairplay.statistics.entity.reservation.EventTicketStatistics;
 
 import java.time.LocalDate;
 import java.util.List;
